requirements.txt
pytesseract
Pillow
requests
jupyter
pandas
openpyxl